requirements.txt
pytest>=7.0
python-dotenv>=1.0

# Optional manual-demo extras
# Install local embedder support with:
# Install OpenAI embedder support with:
openai
# Install both optional embedders with:
sentence-transformers>=3.0

# Optional: for bonus tasks
chromadb>=0.4.0      # Real vector database integration
